Add hooks
