+++ src/components/BoardCard/index.tsx
@@ -1,9 +1,5 @@
 'use client';
 
-import { useState } from 'react';
-
-import { toast } from 'react-toastify';
-
 import { MiniProfile } from '..';
 import { LikeButton } from '..';
 
@@ -11,15 +7,10 @@ import * as S from './style';
 
 import * as I from '@/assets';
 import { BOARD_PATH } from '@/constants';
-import { useGetBoardList, usePostLikeCount } from '@/hooks';
+import { useGetBoardList, useOptimisticLike } from '@/hooks';
 import type { BoardInfoType } from '@/types';
 import { ReverseCategoryType } from '@/types';
 import { parseDateString } from '@/utils';
-
-interface LikeContextType {
-  previousLikeCount: number;
-  previousIsLike: boolean;
-}
 
 const BoardCard: React.FC<BoardInfoType> = ({
   id,
@@ -33,39 +24,11 @@ const BoardCard: React.FC<BoardInfoType> = ({
   isLike,
 }) => {
   const { monthDay, time } = parseDateString(createdAt);
-  const [optimisticLikeCount, setOptimisticLikeCount] = useState(likeCount);
-  const [optimisticIsLike, setOptimisticIsLike] = useState(isLike);
 
   const { refetch } = useGetBoardList();
 
-  const { mutate: postMutate } = usePostLikeCount(id, {
-    onMutate: async (): Promise<LikeContextType> => {
-      setOptimisticLikeCount((prev) =>
-        optimisticIsLike ? prev - 1 : prev + 1
-      );
-      setOptimisticIsLike((prev) => !prev);
-
-      return {
-        previousLikeCount: optimisticLikeCount,
-        previousIsLike: optimisticIsLike,
-      };
-    },
-    onError: (err, variables, context) => {
-      const contextData = context as LikeContextType;
-      if (contextData) {
-        setOptimisticLikeCount(contextData.previousLikeCount);
-        setOptimisticIsLike(contextData.previousIsLike);
-      }
-      toast.error('좋아요 업데이트에 실패했습니다.');
-    },
-    onSuccess: () => {
-      refetch();
-    },
-  });
-
-  const uploadLike = () => {
-    postMutate();
-  };
+  const { optimisticLikeCount, optimisticIsLike, uploadLike } =
+    useOptimisticLike(id, likeCount, isLike, refetch);
 
   return (
     <S.ContentBox>
