+++ src/hooks/api/board/index.ts
@@ -2,4 +2,5 @@ export * from './comment';
 export * from './comment/usePostComment';
 export * from './useGetBoardDetail';
 export * from './useGetBoardList';
+export * from './useOptimisticLike';
 export * from './usePostBoardContent';
